Migrate more eventscript code to EventScriptModule
