+++ ScriptedEvents/API/Modules/EventHandlingModule.cs
@@ -495,106 +495,6 @@
             }
         }
 
-        // Reflection: ON config
-        public EventScriptModule ESModule => MainPlugin.GetModule<EventScriptModule>();
-
-        public void OnAnyEvent(string eventName, IExiledEvent ev = null)
-        {
-            // TODO: Temporary workaround
-            // Switch to handling this entirely in the EventScriptModule
-            if (ESModule.CurrentEventData is null || !ESModule.CurrentEventData.TryGetValue(eventName, out List<string> scripts))
-            {
-                return;
-            }
-
-            foreach (string script in scripts)
-            {
-                try
-                {
-                    Script scr = MainPlugin.ScriptModule.ReadScript(script, null);
-
-                    // Add variables based on event.
-                    if (ev is IPlayerEvent playerEvent && playerEvent.Player is not null)
-                    {
-                        scr.AddPlayerVariable("{EVPLAYER}", "The player that is involved with this event.", new[] { playerEvent.Player });
-                    }
-
-                    if (ev is IAttackerEvent attackerEvent && attackerEvent.Attacker is not null)
-                    {
-                        scr.AddPlayerVariable("{EVATTACKER}", "The attacker that is involved with this event.", new[] { attackerEvent.Attacker });
-                    }
-
-                    if (ev is IItemEvent item && item.Item is not null)
-                    {
-                        scr.AddVariable("{EVITEM}", "The Id of the ItemType of the item involved with this event.", item.Item.Base.ItemSerial.ToString());
-                    }
-                    else if (ev is IPickupEvent pickup && pickup.Pickup is not null)
-                    {
-                        scr.AddVariable("{EVITEM}", "The Id of the ItemType of the pickup associated with this event.", pickup.Pickup.Serial.ToString());
-                    }
-
-                    if (ev is BanningEventArgs ban)
-                    {
-                        scr.AddPlayerVariable("{EVPLAYER}", "The ban issuer.", new[] { ban.Player });
-                        scr.AddVariable("{EVREASON}", "The ban reason.", ban.Reason.ToString());
-                        scr.AddVariable("{EVDURATION}", "The ban duration.", ban.Duration.ToString());
-                        scr.AddVariable("{EVTARGET}", "The ban target.", ban.Target.Nickname);
-                    }
-
-                    if (ev is LocalReportingEventArgs rep)
-                    {
-                        scr.AddPlayerVariable("{EVPLAYER}", "The report issuer.", new[] { rep.Player });
-                        scr.AddVariable("{EVREASON}", "The report reason.", rep.Reason.ToString());
-                        scr.AddPlayerVariable("{EVTARGET}", "The report target.", new[] { rep.Target });
-                    }
-
-                    if (ev is KickingEventArgs kick)
-                    {
-                        scr.AddPlayerVariable("{EVPLAYER}", "The issuer.", new[] { kick.Player });
-                        scr.AddVariable("{EVREASON}", "The reason.", kick.Reason.ToString());
-                        scr.AddPlayerVariable("{EVTARGET}", "The target.", new[] { kick.Target });
-                    }
-
-                    if (ev is HurtingEventArgs hurting)
-                    {
-                        scr.AddVariable("{EVAMMOUNT}", "The amount of damage dealt.", hurting.Amount.ToString());
-                    }
-
-                    if (ev is IDoorEvent door && door.Door is not null)
-                    {
-                        scr.AddVariable("{EVDOOR}", "The door type.", door.Door.Type.ToString());
-                    }
-
-                    MainPlugin.ScriptModule.RunScript(scr);
-                }
-                catch (DisabledScriptException)
-                {
-                    Log.Warn(ErrorGen.Get(ErrorCode.On_DisabledScript, eventName, script));
-                }
-                catch (FileNotFoundException)
-                {
-                    Log.Warn(ErrorGen.Get(ErrorCode.On_NotFoundScript, eventName, script));
-                }
-                catch (Exception ex)
-                {
-                    Log.Warn(ErrorGen.Get(ErrorCode.On_UnknownError, eventName) + $": {ex}");
-                }
-            }
-        }
-
-        public static void OnArgumentedEvent<T>(T ev)
-            where T : IExiledEvent
-        {
-            Type evType = typeof(T);
-            string evName = evType.Name.Replace("EventArgs", string.Empty);
-            MainPlugin.Handlers.OnAnyEvent(evName, ev);
-        }
-
-        public static void OnNonArgumentedEvent()
-        {
-            MainPlugin.Handlers.OnAnyEvent(new StackFrame(2).GetMethod().Name);
-        }
-
         // Infection
         public void OnDied(DiedEventArgs ev)
         {
